refactor(observability): share OTLP signal helpers
